services/streamlit/requirements-test.txt
pytest-cov>=4.1.0,<5
flake8>=6.1.0
black>=23.12.1
requests-mock>=1.11.0,<2
